Move turn component catalog indexers into shared concepts module.
Fleet composition and scores inference both need host catalog lookups;
concepts is the canonical home instead of fleet importing from inference.
