+++ packages/api/api/analytics/military_score_inference/ship_inventory.py
@@ -12,25 +12,15 @@ from api.analytics.military_score_inference.ship_build_combos import (
 from api.analytics.military_score_inference.ship_build_scoring import (
     ship_build_has_zero_military_score,
 )
-from api.models.components import Beam, Engine, Hull, Torpedo
+from api.concepts.turn_component_catalog import (
+    beams_by_id,
+    engines_by_id,
+    hulls_by_id,
+    torpedos_by_id,
+)
+from api.models.components import Hull
 from api.models.game import TurnInfo
 from api.models.ship import Ship
-
-
-def hulls_by_id(turn: TurnInfo) -> dict[int, Hull]:
-    return {hull.id: hull for hull in turn.hulls}
-
-
-def engines_by_id(turn: TurnInfo) -> dict[int, Engine]:
-    return {engine.id: engine for engine in turn.engines}
-
-
-def beams_by_id(turn: TurnInfo) -> dict[int, Beam]:
-    return {beam.id: beam for beam in turn.beams}
-
-
-def torpedos_by_id(turn: TurnInfo) -> dict[int, Torpedo]:
-    return {torp.id: torp for torp in turn.torpedos}
 
 
 def owned_active_ships(turn: TurnInfo, player_id: int) -> list[Ship]:
